added more bosses, fixed a bug

+++ AuthQuest/Assets/Scripts/HealthBarScript.cs
@@ -60,7 +60,7 @@ public class HealthBarScript : MonoBehaviour
         }
     }
 
-    public void HealthRegenerate(float regenValue)
+    public void HealthRegenerate(int regenValue)
     {
         float regenValueNormalized = ((float)regenValue)/((float)(player.GetComponent<PlayerStats>().Health));
         if((currentScale + regenValueNormalized) > 1f)
@@ -70,6 +70,9 @@ public class HealthBarScript : MonoBehaviour
         else
         {
         bar.localScale = new Vector3(currentScale + regenValueNormalized, 1f);
+        //Debug.Log("regenValue = " + regenValue);
+        //Debug.Log("(float)regenValue = " + (float)regenValue);
+        //Debug.Log("health restored " + regenValueNormalized);
         }
         currentScale = bar.localScale.x;
         if (currentScale <= .3)
